@@ -1,6 +1,6 @@
 {
   "name": "thor-devkit",
-  "version": "1.0.0",
+  "version": "1.1.0-beta1",
   "description": "Typescript library to aid DApp development on VeChain Thor",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
@@ -56,28 +56,23 @@
   },
   "homepage": "https://github.com/vechain/thor-devkit.js#readme",
   "devDependencies": {
-    "@types/bip39": "^2.4.0",
     "@types/chai": "^4.1.3",
+    "@types/elliptic": "^6.4.9",
     "@types/mocha": "^5.2.1",
     "@types/node": "^10.5.2",
     "chai": "^4.1.2",
     "coveralls": "^3.0.1",
     "mocha": "^5.2.0",
-    "nyc": "^12.0.2",
-    "ts-node": "^6.1.0",
+    "nyc": "^14.1.1",
+    "ts-node": "^8.3.0",
     "tslint": "^5.11.0",
-    "typescript": "^2.9.1"
+    "typescript": "^3.6.3"
   },
   "dependencies": {
     "@vechain/ethers": "^4.0.27-2",
     "bignumber.js": "^7.2.1",
-    "bip39": "^2.5.0",
     "blakejs": "^1.1.0",
     "fast-json-stable-stringify": "^2.0.0",
-    "hdkey": "^1.0.0",
-    "keccak": "^1.4.0",
-    "keythereum": "^1.0.4",
-    "rlp": "^2.0.0",
-    "secp256k1": "3.5.0"
+    "rlp": "^2.0.0"
   }
 }
